MappedScop::reductionMapSchedule drop elements from list
The function isl_multi_union_pw_aff_drop_dims will not be exposed
in the mainline C++ interface.
A similar function that drops base expression in the multi expression
may get added later and this function could then be used here.
For now, use slightly more verbose code to perform
the dropping of elements on a list extracted from the multi expression.